# Conflicts:
#	CHANGELOG/README.md
#	CHANGELOG/zh_CN/CHANGELOG-2.0.md
#	src/backend/ci/core/project/biz-project/src/main/kotlin/com/tencent/devops/project/service/impl/AbsProjectServiceImpl.kt

+++ src/backend/ci/core/stream/biz-stream/src/main/kotlin/com/tencent/devops/stream/dao/GitRequestEventDao.kt
@@ -367,7 +367,7 @@ class GitRequestEventDao {
      */
     fun getRequestsById(
         dslContext: DSLContext,
-        requestIds: Set<Int>,
+        requestIds: Set<Long>,
         hasEvent: Boolean
     ): List<GitRequestEvent> {
         with(TGitRequestEvent.T_GIT_REQUEST_EVENT) {
